+++ src/app/helpers/tests/debugUnblock.test.js
@@ -93,16 +93,17 @@ describe('debugUnblock', () => {
     expect(logMock.mock.calls[0][0].replace(/\r\n/g, '\n')).toEqual(`-- UNBLOCKING:
 method              id: 1  wait: 43       
 -- TREE:
-                    id: 7  wait: all       Children: [3,4,8]
-                    id: 8  wait: all       Children: [5,6]
-root                id: 0  wait: generator Children: [1,2,3,4,5,6,7,9]
-calledMethod        id: 9  wait: generator Children: [10]
+calledMethod        id: 10 wait: generator Children: [11]
+                    id: 7  wait: all       Children: [1,2,8,10]
+                    id: 8  wait: all       Children: [3,4,9]
+                    id: 9  wait: all       Children: [5,6]
+root                id: 0  wait: generator Children: [7]
 method              id: 1  wait: 43        Children: []
 methodNested        id: 6  wait: 53        Children: []
 method              id: 4  wait: 56        Children: []
 method              id: 3  wait: 65        Children: []
 method              id: 5  wait: 97        Children: []
-method              id: 10 wait: 110       Children: []
+method              id: 11 wait: 110       Children: []
 method              id: 2  wait: true      Children: []
 `.replace(/\r\n/g, '\n'));
 
@@ -110,19 +111,20 @@ method              id: 2  wait: true      Children: []
     expect(logMock.mock.calls[2][0].replace(/\r\n/g, '\n')).toEqual(`-- UNBLOCKING:
 method              id: 4  wait: 51       
 -- TREE:
-                    id: 7  wait: all       Children: [3,4,8]
-                    id: 8  wait: all       Children: [5,6]
-root                id: 0  wait: generator Children: [2,3,4,5,6,7,9]
-calledMethod        id: 9  wait: generator Children: [10]
-deeplyNestedMethodWithVeryLongNameid: 13 wait: generator Children: [14]
-methodNested        id: 6  wait: 48        Children: [11,12,13]
+calledMethod        id: 10 wait: generator Children: [11]
+deeplyNestedMethodWithVeryLongNameid: 14 wait: generator Children: [15]
+                    id: 7  wait: all       Children: [2,8,10]
+                    id: 8  wait: all       Children: [3,4,9]
+                    id: 9  wait: all       Children: [5,6]
+root                id: 0  wait: generator Children: [7]
+methodNested        id: 6  wait: 48        Children: [14]
 method              id: 4  wait: 51        Children: []
 method              id: 3  wait: 60        Children: []
-method              id: 12 wait: 78        Children: []
-method              id: 14 wait: 90        Children: []
+method              id: 13 wait: 78        Children: []
+method              id: 15 wait: 90        Children: []
 method              id: 5  wait: 92        Children: []
-method              id: 10 wait: 105       Children: []
-method              id: 11 wait: 197       Children: []
+method              id: 11 wait: 105       Children: []
+method              id: 12 wait: 197       Children: []
 method              id: 2  wait: true      Children: []
 `.replace(/\r\n/g, '\n'));
 
@@ -130,7 +132,8 @@ method              id: 2  wait: true      Children: []
     expect(logMock.mock.calls[11][0].replace(/\r\n/g, '\n')).toEqual(`-- UNBLOCKING:
 method              id: 2  wait: true     
 -- TREE:
-root                id: 0  wait: generator Children: [2]
+root                id: 0  wait: generator Children: [7]
+                    id: 7  wait: all       Children: [2]
 method              id: 2  wait: true      Children: []
 `.replace(/\r\n/g, '\n'));
     logMock.mockRestore();